Move arg parsing to new file
